@@ -1,15 +1,17 @@
 {
   "name": "sample-nodejs",
   "version": "0.0.1",
+  "type": "module",
   "dependencies": {
     "body-parser": "1.19.0",
     "express": "4.17.1"
   },
   "scripts": {
-    "start": "node web.js"
+    "start": "node --experimental-specifier-resolution=node web.js",
+    "dev": "NODE_ENV=development nodemon --experimental-specifier-resolution=node index.js | pino-pretty"
   },
   "engines": {
-    "node": ">= 12.0.0"
+    "node": "14"
   },
   "devDependencies": {
     "nodemon": "^2.0.18"
