@@ -1,17 +1,10 @@
-const express = require('express');
+import express from 'express';
+import bodyParser from 'body-parser';
+import router from './routes';
+
 const app = express();
-const bodyParser = require('body-parser');
 
 app.use(bodyParser.json());
-
-app.get('/', function (req, res) {
-  res.send('Let the battle begin!');
-});
-
-app.post('/', function (req, res) {
-  console.log(req.body);
-  const moves = ['F', 'T', 'L', 'R'];
-  res.send(moves[Math.floor(Math.random() * moves.length)]);
-});
+app.use('/', router);
 
 app.listen(process.env.PORT || 8080);
